+++ CHANGELOG.md
@@ -18,6 +18,7 @@ dependencies:
 1. **OPTIMIZATION** Prisma Dart client generator outputs more user-friendly error messages that help you know where the problem lies.
 2. **REFACTOR** Adjust src code storage structure.
 3. **FIX** Rethrow start error
+4. **FIX** Fix input types required multiple fields.
 
 # Prisma Dart client v4.0.0-alpha.3